requirements.txt

streamlit>=1.36,<2
praw>=7.7,<8
pandas>=2.2,<3
numpy>=1.26,<2
yfinance>=0.2.40,<0.3
vaderSentiment>=3.3.2,<4
python-dotenv>=1.0,<2
statsmodels>=0.14,<0.15
